Handle negotiation exception. (#8)
* Handle negotiation exception.

If connection negotiation failes in __init__() the ThreadPoolExecutor
thread is left ofer and the python process does not quit on the first
CTRL-C and as __init__() fails there is not instance to call close().

This PR handles the exception and calls close() if negotiation fails.

* Implement reconnect on connection failure.

This PR implements reconnect on connection failure.