Resendbuffer (#12)
* Add timeout to RelpTCPClient.close().

When calling RelpTCPClient.close() shortly after the last message was
sent its likely to get a "txnr does not found" error because the ACK was
not received yet. Using a timeout value of e.g. 1 fixes this issue.

* Update relppy/client.py

Co-authored-by: Watanabe Takashi <wtnb75@gmail.com>

* Import missing concurrent.

* Improve resend buffer handling.

Add resend_interval argument and use it to process resend-buffer not only
if it reaches resend_bufsize but also on the given interval. Start an
additional ThreadPoolExecutor for self.sender() to process command/data
to send and to process resend buffer. This addresses #11.

* Fix log message (wrong line)

* Notify sender thread about close.

* Update relppy/client.py

Co-authored-by: Watanabe Takashi <wtnb75@gmail.com>

* Resend only messages that reached resend_interval.2

In the previous implementation resend was resending all messages in
resendbuf. Now it only resends if the message was buffered for
self.resend_interval.

---------

Co-authored-by: Watanabe Takashi <wtnb75@gmail.com>